@@ -4,12 +4,12 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve ",
-    "build": "ng build --prod --aot=true",
+    "build": "ng build --configuration production --aot=true",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
     "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && yarn install && yarn start",
-    "ngh:build": "ng build --prod",
+    "ngh:build": "ng build --configuration production",
     "ngh:deploy": "ng deploy --branch=master --base-href=/app-spa-repesc/ --repo=https://github.com/fean-developer/app-spa-repesc.git --no-silent"
   },
   "engines": {
@@ -19,21 +19,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "~0.1102.10",
-    "@angular/animations": "~11.2.11",
+    "@angular-devkit/build-angular": "~12.2.16",
+    "@angular/animations": "~12.2.16",
     "@angular/cdk": "^11.2.11",
-    "@angular/cli": "~11.2.10",
-    "@angular/common": "~11.2.11",
-    "@angular/compiler": "~11.2.11",
-    "@angular/compiler-cli": "~11.2.11",
-    "@angular/core": "~11.2.11",
-    "@angular/forms": "~11.2.11",
+    "@angular/cli": "~12.2.16",
+    "@angular/common": "~12.2.16",
+    "@angular/compiler": "~12.2.16",
+    "@angular/compiler-cli": "~12.2.16",
+    "@angular/core": "~12.2.16",
+    "@angular/forms": "~12.2.16",
     "@angular/material": "^11.2.11",
-    "@angular/platform-browser": "~11.2.11",
-    "@angular/platform-browser-dynamic": "~11.2.11",
-    "@angular/router": "~11.2.11",
-    "@ngrx/store": "11",
-    "@ngrx/store-devtools": "11",
+    "@angular/platform-browser": "~12.2.16",
+    "@angular/platform-browser-dynamic": "~12.2.16",
+    "@angular/router": "~12.2.16",
+    "@ngrx/store": "12",
+    "@ngrx/store-devtools": "12",
     "@popperjs/core": "^2.9.2",
     "bootstrap": "^5.0.0-beta3",
     "jsencrypt": "^3.2.1",
@@ -42,7 +42,7 @@
     "ngx-mask": "^11.1.5",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.3"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
     "@types/jasmine": "~3.6.0",
@@ -51,7 +51,7 @@
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.1.0",
+    "karma": "~6.3.16",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
@@ -59,6 +59,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.1.5"
+    "typescript": "~4.3.5"
   }
 }
