@@ -6,6 +6,6 @@ export const selectState = (state: AppState) => state
 
 export const selecRepescs = createSelector(
     selectState,
-    (state)  => { console.log('STATE SELETOR:', state); return state}
+    (state)  =>  state
 )
 
